src/requirements.txt
aiokafka==0.12.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
build==1.2.2.post1
click==8.1.7
confluent-kafka==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pip-tools==7.4.1
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
